.history/requirements_20251105141638.txt
Flask==2.3.0
Werkzeug==2.3.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
